+++ src/shared/ui/Button.tsx
@@ -1,0 +1,19 @@
+import type { VariantProps } from 'class-variance-authority';
+import { type ButtonHTMLAttributes, forwardRef } from 'react';
+
+import { buttonVariants } from '../config';
+
+/**
+ * 버튼 컴포넌트
+ */
+interface ButtonProps
+  extends ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  className?: string;
+}
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, ...props }, ref) => (
+    <button className={buttonVariants({ variant, size, className })} ref={ref} {...props} />
+  ),
+);
+Button.displayName = 'Button';
