+++ src/shared/ui/Input.tsx
@@ -1,0 +1,19 @@
+import { type InputHTMLAttributes, forwardRef } from 'react';
+
+/**
+ * 입력 컴포넌트
+ */
+interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
+  className?: string;
+}
+export const Input = forwardRef<HTMLInputElement, InputProps>(
+  ({ className, type, ...props }, ref) => (
+    <input
+      type={type}
+      className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      ref={ref}
+      {...props}
+    />
+  ),
+);
+Input.displayName = 'Input';
